pyproject.toml
[tool.poetry]
name = "sean-gpt"
version = "0.1.0"
description = ""
authors = ["Sean F <fitzgerald.sean.t@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.36.2"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pydantic = "^2.5.2"
openai = "^1.5.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
twilio = "^8.11.0"
httpx = "^0.26.0"
pytest = "^7.4.3"
docker = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
